requirements.txt
python-telegram-bot==13.15
ccxt==3.0.75
requests==2.28.2
python-dotenv==1.0.0
schedule==1.2.0
